Move registration reminder logic to a service PORO
